Add jsdoc to transactions utility
