+++ src/utilities/api/transactions.js
@@ -1,0 +1,61 @@
+import getMappedFunction from './functionMapper';
+
+/**
+ * @typedef {Object} Transaction
+ * @property {String} id ID or hash of the transaction.
+ * @property {String} senderAddress
+ * @property {String} recipientAddress
+ * @property {String} amount
+ * @property {String} fee
+ * @property {Number} timestamp
+ * @property {Number} confirmations
+ * @property {Object} extra additional or uncommon properties.
+ * @property {Number} extra.type transaction type for LSK.
+ * @property {String} extra.data transaction data field for LSK.
+ */
+
+/**
+ * @typedef {Object} TransactionsResponse
+ * @property {Array.<Transaction>} data
+ * @property {Object} meta
+ * @property {Number} meta.count total number of transactions available on the API.
+ */
+
+/**
+ * Retrieves transactions for related token with respect to the given payload.
+ * @param {String} tokenType
+ * @param {Object} data
+ * @param {String} data.id ID or hash of a specific transaction.
+ * @param {String} data.address
+ * @param {Number} data.limit
+ * @param {Number} data.offset
+ * @returns {Promise<TransactionsResponse>}
+ */
+const get = (tokenType, data) => getMappedFunction(tokenType, 'transactions.get')(data);
+
+/**
+ * Creates a raw, ready-to-broadcast transaction data with given payload.
+ * @param {String} tokenType
+ * @param {Object} data
+ * @param {String} data.passphrase
+ * @param {String} data.recipientAddress
+ * @param {String} data.amount
+ * @param {String} data.secondPassphrase required if registered, only used for LSK.
+ * @param {String} data.data custom data field, only used for LSK.
+ * @returns {Promise}
+ */
+const create = (tokenType, data) => getMappedFunction(tokenType, 'transactions.create')(data);
+
+/**
+ * Broadcasts a transaction to the specified token's blockchain.
+ * @param {String} tokenType
+ * @param {Object} transaction
+ * @returns {Promise}
+ */
+const broadcast = (tokenType, transaction) => getMappedFunction(tokenType, 'transactions.broadcast')(transaction);
+
+export default {
+  get,
+  create,
+  broadcast,
+};
